PhotoSnap/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
distlib==0.3.8
Django==5.0.3
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-environ==0.11.2
django-filter==24.1
django-rest-auth-forked==0.9.5
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
environ==1.0
filelock==3.13.1
Markdown==3.6
mysql-connector-python==8.3.0
mysqlclient==2.2.4
pbr==6.0.0
pipenv==2023.12.1
platformdirs==4.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
stevedore==5.2.0
typing_extensions==4.10.0
